requirements.txt
django
django-debug-toolbar
Pillow
django-ckeditor

